requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==4.2.9
django-debug-toolbar==4.4.6
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
idna==3.7
kombu==5.4.0
oauthlib==3.2.2
pillow==10.4.0
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
text-unidecode==1.3
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
